refactor: extract infographic timeline composable
